Add useCalendar hook
